+++ tests/core/test_util.py
@@ -19,6 +19,7 @@ from pythainlp.util import (
     bahttext,
     collate,
     convert_years,
+    count_thai,
     count_thai_chars,
     countthai,
     dict_trie,
@@ -29,6 +30,8 @@ from pythainlp.util import (
     expand_maiyamok,
     find_keyword,
     ipa_to_rtgs,
+    is_thai,
+    is_thai_char,
     isthai,
     isthaichar,
     longest_common_subsequence,
@@ -716,6 +719,11 @@ class UtilTestCase(unittest.TestCase):
         self.assertFalse(isthaichar("a"))
         self.assertFalse(isthaichar("0"))
 
+    def test_is_thai_char(self):
+        self.assertTrue(is_thai_char("ก"))
+        self.assertFalse(is_thai_char("a"))
+        self.assertFalse(is_thai_char("0"))
+
     def test_isthai(self):
         self.assertTrue(isthai("ไทย"))
         self.assertTrue(isthai("ต.ค."))
@@ -723,6 +731,23 @@ class UtilTestCase(unittest.TestCase):
         self.assertFalse(isthai("ไทย0"))
         self.assertFalse(isthai("(ต.ค.)"))
         self.assertFalse(isthai("ต.ค.", ignore_chars=None))
+
+    def test_is_thai(self):
+        self.assertTrue(is_thai("ไทย"))
+        self.assertTrue(is_thai("ต.ค."))
+        self.assertTrue(is_thai("(ต.ค.)", ignore_chars=".()"))
+        self.assertFalse(is_thai("ไทย0"))
+        self.assertFalse(is_thai("(ต.ค.)"))
+        self.assertFalse(is_thai("ต.ค.", ignore_chars=None))
+
+    def test_count_thai(self):
+        self.assertEqual(count_thai(""), 0.0)
+        self.assertEqual(count_thai("123"), 0.0)
+        self.assertEqual(count_thai("1 2 3"), 0.0)
+        self.assertEqual(count_thai("ประเทศไทย"), 100.0)
+        self.assertEqual(count_thai("โรค COVID-19"), 37.5)
+        self.assertEqual(count_thai("(กกต.)", ".()"), 100.0)
+        self.assertEqual(count_thai("(กกต.)", None), 50.0)
 
     def test_display_thai_char(self):
         self.assertEqual(display_thai_char("้"), "_้")